Plot method for simList objects

Extends Plot for simList objects.

# S4 method for simList
Plot(..., new = FALSE, addTo = NULL, gp = gpar(),
  gpText = gpar(), gpAxis = gpar(), axes = FALSE, speedup = 1,
  size = 5, cols = NULL, col = NULL, zoomExtent = NULL,
  visualSqueeze = NULL, legend = TRUE, legendRange = NULL,
  legendText = NULL, pch = 19, title = NULL, na.color = "#FFFFFF00",
  zero.color = NULL, length = NULL, arr = NULL, plotFn = "plot")

Arguments

...
A combination of spatialObjects or non-spatial objects. For many object classes, there are specific Plot methods. Where there are no specific ones, the base plotting will be used internally. This means that for objects with no specific Plot methods, many arguments, such as addTo, will not work. See details.

new
Logical. If TRUE, then the previous named plot area is wiped and a new one made; if FALSE, then the ... plots will be added to the current device, adding or rearranging the plot layout as necessary. Default is FALSE. This currently works best if there is only one object being plotted in a given Plot call. However, it is possible to pass a list of logicals to this, matching the length of the ... objects. Use clearPlot to clear the whole plotting device.

addTo
Character vector, with same length as .... This is for overplotting, when the overplot is not to occur on the plot with the same name, such as plotting a SpatialPoints* object on a RasterLayer.

gp
A gpar object, created by gpar function, to change plotting parameters (see grid package).

gpText
A gpar object for the title text. Default gpar(col = "black").

gpAxis
A gpar object for the axes. Default gpar(col = "black").

axes
Logical or "L", representing the left and bottom axes, over all plots.

speedup
Numeric. The factor by which the number of pixels is divided by to plot rasters. See Details.

size
Numeric. The size, in points, for SpatialPoints symbols, if using a scalable symbol.

cols
(also col) Character vector or list of character vectors of colours. See details.

col
(also cols) Alternative to cols to be consistent with plot. cols takes precedence, if both are provided.

zoomExtent
An Extent object. Supplying a single extent that is smaller than the rasters will call a crop statement before plotting. Defaults to NULL. This occurs after any downsampling of rasters, so it may produce very pixelated maps.

visualSqueeze
Numeric. The proportion of the white space to be used for plots. Default is 0.75.

legend
Logical indicating whether a legend should be drawn. Default is TRUE.

legendRange
Numeric vector giving values that, representing the lower and upper bounds of a legend (i.e., 1:10 or c(1,10) will give same result) that will override the data bounds contained within the grobToPlot.

legendText
Character vector of legend value labels. Defaults to NULL, which results in a pretty numeric representation. If Raster* has a Raster Attribute Table (rat; see raster package), this will be used by default. Currently, only a single vector is accepted. The length of this must match the length of the legend, so this is mostly useful for discrete-valued rasters.

pch
see ?par.

title
Logical or character string. If logical, it indicates whether to print the object name as the title above the plot. If a character string, it will print this above the plot. NOTE: the object name is used with addTo, not the title. Default NULL, which means print the object name as title, if no other already exists on the plot, in which case, keep the previous title.

na.color
Character string indicating the color for NA values. Default transparent.

zero.color
Character string indicating the color for zero values, when zero is the minimum value, otherwise, zero is treated as any other color. Default transparent.

length
Numeric. Optional length, in inches, of the arrow head.

arr
A vector of length 2 indicating a desired arrangement of plot areas indicating number of rows, number of columns. Default NULL, meaning let Plot function do it automatically.

plotFn
An optional function name to do the plotting internally, e.g., "barplot" to get a barplot() call. Default "plot".

Details

Plot for simList class objects

See Plot. This method strips out stuff from a simList class object that would make it otherwise not reproducibly digestible between sessions, operating systems, or machines. This will likely still not allow identical digest results across R versions.

See also