Plot wrapper intended for use in a SpaDES module

This is a single function call that allows a module to change which format in which the plots will occur. Specifically, the two primary formats would be to "screen" or to disk as an image file, such as "png". THIS CURRENTLY ONLY WORKS CORRECTLY WITH ggplot2 objects that can be saved. It uses Plot internally, so individual plots may be rearranged. This function requires at least 2 things: a plotting function and data for that plot function. See below and examples.

Plots(
  data,
  fn,
  filename,
  types = quote(params(sim)[[currentModule(sim)]]$.plots),
  path = quote(file.path(outputPath(sim), "figures")),
  .plotInitialTime = quote(params(sim)[[currentModule(sim)]]$.plotInitialTime),
  ggsaveArgs = list(),
  usePlot = TRUE,
  deviceArgs = list(),
  ...
)

Arguments

data
An arbitrary data object. It should be used inside the Plots function, and should contain all the data required for the inner plotting

fn
An arbitrary plotting function.

filename
A name that will be the base for the files that will be saved, i.e, do not supply the file extension, as this will be determined based on types. If a user provides this as an absolute path, it will override the path argument.

types
Character vector, zero or more of types. See below.

path
Currently a single path for the saved objects on disk. If filename is supplied as an absolute path, path will be set to dirname(filename), overriding this argument value.

.plotInitialTime
A numeric. If NA then no visual on screen. Anything else will have visuals plotted to screen device. This is here for backwards compatibility. A developer should set in the module to the intended initial plot time and leave it.

ggsaveArgs
An optional list of arguments passed to ggplot2::ggsave

usePlot
Logical. If TRUE, the default, then the plot will occur with quickPlot::Plot, so it will be arranged with previously existing plots.

deviceArgs
An optional list of arguments passed to one of png, pdf, tiff, bmp, or jgeg. This is useful when the plotting function is not creating a ggplot object.

...
Anything needed by fn

Details

type

"screen" -- Will plot to the current device, normally a plot window

"object" -- Will save the plot object, e.g., ggplot object

"raw" -- Will save the raw data prior to plotting, e.g., the data argument

"png" -- or any other type save-able with ggsave

Note

THIS IS STILL EXPERIMENTAL and could change in the next release.

Plots now has experimental support for "just a Plot call", but with types specified. See example. The devices to save on disk will have some different behaviours to the screen representation, since "wiping" an individual plot on a device doesn't exist for a file device.

This offers up to 4 different actions for a given plot:

To screen device

To disk as raw data

To disk as a saved plot object

To disk as a .png or other image file, e.g., .pdf

To turn off plotting both to screen and disk, set both .plotInititalTime = NA and .plots = NA or any other value that will not trigger a TRUE with a grepl with the types argument (e.g., "" will omit all saving).

Examples

if (FALSE) {
# Note: if this is used inside a SpaDES module, do not define this
#  function inside another function. Put it outside in a normal
#  module script. It will cause a memory leak, otherwise.
if (!require("ggplot2")) stop("please install ggplot2")
fn <- function(d)
  ggplot(d, aes(a)) +
  geom_histogram()
sim <- simInit()
sim$something <- data.frame(a = sample(1:10, replace = TRUE))

Plots(data = sim$something, fn = fn,
      types = c("png"),
      path = file.path("figures"),
      filename = tempfile(),
      .plotInitialTime = 1
      )

# plot to active device and to png
Plots(data = sim$something, fn = fn,
      types = c("png", "screen"),
      path = file.path("figures"),
      filename = tempfile(),
      .plotInitialTime = 1
      )

# Can also be used like quickPlot::Plot, but with control over output type
r <- raster::raster(extent(0,10,0,10), vals = sample(1:3, size = 100, replace = TRUE))
Plots(r, types = c("screen", "png"), deviceArgs = list(width = 700, height = 500))

} # end of dontrun