Simulation checkpoints

Save and reload the current state of the simulation, including the state of the random number generator, by scheduling checkpoint events.

checkpointFile(sim)

# S4 method for simList
checkpointFile(sim)

checkpointFile(sim) <- value

# S4 method for simList
checkpointFile(sim) <- value

checkpointInterval(sim)

# S4 method for simList
checkpointInterval(sim)

checkpointInterval(sim) <- value

# S4 method for simList
checkpointInterval(sim) <- value

doEvent.checkpoint(sim, eventTime, eventType, debug = FALSE)

checkpointLoad(file)

.checkpointSave(sim, file)

Arguments

sim

A simList simulation object.

value

The parameter value to be set (in the corresponding module and param).

eventTime

A numeric specifying the time of the next event.

eventType

A character string specifying the type of event: one of either "init", "load", or "save".

debug

Optional logical flag determines whether sim debug info will be printed (default debug = FALSE).

file

The checkpoint file.

Value

Returns the modified simList object.

Note

Checkpoint files are intended to be used locally, and do not invoke the simulation archiving tools to bundle and subsequently extract simulation files (e.g., file-backed rasters).

RNG save code adapted from: http://www.cookbook-r.com/Numbers/Saving_the_state_of_the_random_number_generator/ and https://stackoverflow.com/q/13997444/1380598

See also

Author

Alex Chubaty