Methods for .wrap and .unwrap

Methods for .wrap and .unwrap

# S3 method for class 'simList'
.wrap(
  obj,
  cachePath,
  preDigest,
  drv = getOption("reproducible.drv", NULL),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  outputObjects = NULL,
  ...
)

# S3 method for class '.simDeps'
.wrap(obj, ...)

# S3 method for class '.simDeps'
.unwrap(obj, ...)

# S3 method for class '.moduleDeps'
.wrap(obj, ...)

# S3 method for class '.moduleDeps'
.unwrap(obj, ...)

# S3 method for class 'simList'
.unwrap(
  obj,
  cachePath,
  cacheId,
  drv = getOption("reproducible.drv", NULL),
  conn = getOption("reproducible.conn", NULL),
  ...
)

Arguments

obj

Any arbitrary R object.

cachePath

A repository used for storing cached objects. This is optional if Cache is used inside a SpaDES module.

preDigest

The list of preDigest that comes from CacheDigest of an object

drv

an object that inherits from DBIDriver, or an existing DBIConnection object (in order to clone an existing connection).

conn

A DBIConnection object, as returned by dbConnect().

verbose

Numeric, -1 silent (where possible), 0 being very quiet, 1 showing more messaging, 2 being more messaging, etc. Default is 1. Above 3 will output much more information about the internals of Caching, which may help diagnose Caching challenges. Can set globally with an option, e.g., options('reproducible.verbose' = 0) to reduce to minimal

outputObjects

Optional character vector indicating which objects to return. This is only relevant for list, environment (or similar) objects

...

Other arguments. Can be in the form of tagKey = tagValue, such as, class = "numeric" to find all entries that are numerics in the cache. Note: the special cases of cacheId and fun have their own named arguments in these functions. Also can be regexp = xx, where xx is TRUE if the user is passing a regular expression. Otherwise, userTags will need to be exact matches. Default is missing, which is the same as TRUE. If there are errors due to regular expression problem, try FALSE. For cc, it is passed to clearCache, e.g., ask, userTags. For showCache, it can also be sorted = FALSE to return the object unsorted.

cacheId

An optional character vector describing the cacheIds to extract. Only entries with this/these cacheIds will be returned. If useDBI(FALSE), this will also be dramatically faster than using userTags, for a large cache.

Value

The same object as passed into the function, but dealt with so that it can be saved to disk.