Define a parameter used in a module

Used to specify a parameter's name, value, and set a default.

defineParameter(name, class, default, min, max, desc)

# S4 method for character,character,ANY,ANY,ANY,character
defineParameter(name,
  class, default, min, max, desc)

  # S4 method for character,character,ANY,missing,missing,character
defineParameter(name,
  class, default, desc)

# S4 method for missing,missing,missing,missing,missing,missing
defineParameter()

Arguments

name
Character string giving the parameter name.

class
Character string giving the parameter class.

default
The default value to use when none is specified by the user. Non-standard evaluation is used for the expression.

min
With max, used to define a suitable range of values. Non-standard evaluation is used for the expression.

max
With min, used to define a suitable range of values. Non-standard evaluation is used for the expression.

desc
Text string providing a brief description of the parameter.

Value

data.frame

Note

Be sure to use the correct NA type: logical (NA), integer (NA_integer_), real (NA_real_), complex (NA_complex_), or character (NA_character_). See NA.

Examples

# NOT RUN {
parameters = rbind(
  defineParameter("lambda", "numeric", 1.23, desc = "intrinsic rate of increase"),
  defineParameter("P", "numeric", 0.2, 0, 1, "probability of attack")
)

# }