Simulation event diagram

Create a Gantt Chart representing the events in a completed simulation. This event diagram is constructed using the completed event list To change the number of events shown, provide an n argument.

eventDiagram(sim, n, startDate, ...)

# S4 method for simList,numeric,character
eventDiagram(sim, n, startDate, ...)

# S4 method for simList,missing,character
eventDiagram(sim, n, startDate, ...)

# S4 method for simList,missing,missing
eventDiagram(sim, n, startDate, ...)

Arguments

sim
A simList object (typically corresponding to a completed simulation).

n
The number of most recently completed events to plot.

startDate
A character representation of date in YYYY-MM-DD format.

...
Additional arguments passed to mermaid. Useful for specifying height and width.

Value

Plots an event diagram as Gantt Chart, invisibly returning a mermaid object.

Details

Simulation time is presented on the x-axis, starting at date 'startDate'. Each module appears in a color-coded row, within which each event for that module is displayed corresponding to the sequence of events for that module. Note that only the start time of the event is meaningful is these figures: the width of the bar associated with a particular module's event DOES NOT correspond to an event's "duration".

Based on this StackOverflow answer: http://stackoverflow.com/a/29999300/1380598.

Note

A red vertical line corresponding to the current date may appear on the figure. This is useful for Gantt Charts generally but can be considered a 'bug' here.

See also

mermaid.