Define an input object that the module expects.

Used to specify an input object's name, class, description, source url and other specifications.

expectsInput(objectName, objectClass, desc, sourceURL, ...)

# S4 method for ANY,ANY,ANY,ANY
expectsInput(objectName, objectClass, desc,
  sourceURL, ...)

# S4 method for character,character,character,character
expectsInput(objectName,
  objectClass, desc, sourceURL, ...)

# S4 method for character,character,character,missing
expectsInput(objectName,
  objectClass, desc, sourceURL, ...)

Arguments

objectName
Character string to define the input object's name.

objectClass
Character string to specify the input object's class.

desc
Text string providing a brief description of the input object.

sourceURL
Character string to specify an URL to reach the input object, default is NA.

...
Other specifications of the input object.

Value

A data.frame suitable to be passed to inputObjects in a module's metadata.

Examples

inputObjects <- dplyr::bind_rows(
  expectsInput(objectName = "inputObject1", objectClass = "character",
               desc = "this is for example", sourceURL = "not available"),
  expectsInput(objectName = "inputObject2", objectClass = "numeric",
               desc = "this is for example", sourceURL = "not available",
               otherInformation = "I am the second input object")
)