Estimate memory used with system("ps")

This will give a slightly different estimate than pryr::mem_used, which uses gc() internally. The purpose of this function is to allow continuous monitoring, external to the R session. Normally, this is run in a different session.

This will only work if the user has specified before running the spades call, set the interval, in seconds, that ps is run. E.g., options("spades.memoryUseInterval" = 0.5), will assess memory use every 0.5 seconds. The default is 0, meaning no interval, "off".

memoryUseThisSession(thisPid)

memoryUse(sim, max = TRUE)

Arguments

thisPid

Numeric or integer, the PID of the process. If omitted, it will be found with Sys.getpid().

sim

A completed simList

max

Logical. If TRUE, then it the return value will be summarized by module/event, showing the maximum memory used. If FALSE, then the raw memory used during each event will be shown.

Value

estimated memory use in MiB

data.table summarizing the estimated memory use (in MiB) for each event type, for each module, during the simulation.

Note

The suggested future and future.callr packages must be available.

See also