Parse and extract module metadata

Parse and extract module metadata

moduleMetadata(module, path, sim)

# S4 method for character,character,missing
moduleMetadata(module, path)

# S4 method for character,missing,missing
moduleMetadata(module)

# S4 method for ANY,missing,simList
moduleMetadata(module, sim)

Arguments

module
Character string. Your module's name.

path
Character string specifying the file path to modules directory. Default is to use the spades.modulePath option.

sim
A simList simulation object, generally produced by simInit.

Value

A list of module metadata, matching the structure in defineModule.

See also

defineModule

Examples

path <- system.file("sampleModules", package = "SpaDES.core")
sampleModules <- dir(path)
x <- moduleMetadata(sampleModules[3], path)

# using simList
mySim <- simInit(
   times = list(start = 2000.0, end = 2002.0, timeunit = "year"),
   params = list(
     .globals = list(stackName = "landscape", burnStats = "nPixelsBurned")
   ),
   modules = list("randomLandscapes", "fireSpread", "caribouMovement"),
   paths = list(modulePath = system.file("sampleModules", package = "SpaDES.core"))
)
moduleMetadata(sim = mySim)

#> $randomLandscapes
#> $randomLandscapes$name
#> [1] "randomLandscapes"
#> 
#> $randomLandscapes$description
#> [1] "Generate RasterStack of random maps representative of a forest landscape (DEM, forestAge, forestCover, habitatQuality, percentPine). Requires a global simulation parameter `stackName` be set."
#> 
#> $randomLandscapes$keywords
#> [1] "random map"       "random landscape"
#> 
#> $randomLandscapes$childModules
#> character(0)
#> 
#> $randomLandscapes$authors
#> [1] "Alex M Chubaty <alexander.chubaty@canada.ca> [aut, cre]" 
#> [2] "Eliot J B McIntire <eliot.mcintire@canada.ca> [aut, cre]"
#> 
#> $randomLandscapes$version
#> $randomLandscapes$version$SpaDES.core
#> [1] "0.1.0"
#> 
#> $randomLandscapes$version$SpaDES.tools
#> [1] "0.1.0"
#> 
#> $randomLandscapes$version$randomLandscapes
#> [1] "1.6.0"
#> 
#> 
#> $randomLandscapes$spatialExtent
#> class       : Extent 
#> xmin        : NA 
#> xmax        : NA 
#> ymin        : NA 
#> ymax        : NA 
#> 
#> $randomLandscapes$timeframe
#> [1] NA NA
#> 
#> $randomLandscapes$timeunit
#> [1] "year"
#> 
#> $randomLandscapes$citation
#> list()
#> 
#> $randomLandscapes$documentation
#> list()
#> 
#> $randomLandscapes$reqdPkgs
#> $randomLandscapes$reqdPkgs[[1]]
#> [1] "raster"
#> 
#> $randomLandscapes$reqdPkgs[[2]]
#> [1] "RColorBrewer"
#> 
#> $randomLandscapes$reqdPkgs[[3]]
#> [1] "tkrplot"
#> 
#> $randomLandscapes$reqdPkgs[[4]]
#> [1] "RandomFields"
#> 
#> $randomLandscapes$reqdPkgs[[5]]
#> [1] "SpaDES.tools"
#> 
#> 
#> $randomLandscapes$parameters
#>          paramName paramClass    default        min   max
#> 1            inRAM    logical      FALSE       TRUE FALSE
#> 2               nx    numeric        100         10   500
#> 3               ny    numeric        100         10   500
#> 4        stackName  character  landscape         NA    NA
#> 5 .plotInitialTime    numeric start(sim) start(sim)    NA
#> 6    .plotInterval    numeric         NA         NA    NA
#> 7 .saveInitialTime    numeric         NA         NA    NA
#> 8    .saveInterval    numeric         NA         NA    NA
#> 9        .useCache    logical      FALSE init, plot    NA
#>                                           paramDesc
#> 1            should the raster be stored in memory?
#> 2 size of map (number of pixels) in the x dimension
#> 3 size of map (number of pixels) in the y dimension
#> 4                           name of the RasterStack
#> 5                 time to schedule first plot event
#> 6                 time interval between plot events
#> 7                 time to schedule first save event
#> 8                 time interval between save events
#> 9 should the module result be cached for future use
#> 
#> $randomLandscapes$inputObjects
#>   objectName objectClass desc sourceURL other
#> 1       <NA>        <NA> <NA>      <NA>    NA
#> 
#> $randomLandscapes$outputObjects
#>                             objectName objectClass desc other
#> 1 P(sim, "randomLandscapes")$stackName RasterStack <NA>    NA
#> 
#> 
#> $caribouMovement
#> $caribouMovement$name
#> [1] "caribouMovement"
#> 
#> $caribouMovement$description
#> [1] "Simulate caribou movement via correlated random walk."
#> 
#> $caribouMovement$keywords
#> [1] "caribou"                         "individual based movement model"
#> [3] "correlated random walk"         
#> 
#> $caribouMovement$childModules
#> character(0)
#> 
#> $caribouMovement$authors
#> [1] "Eliot J B McIntire <eliot.mcintire@canada.ca> [aut, cre]"
#> 
#> $caribouMovement$version
#> $caribouMovement$version$SpaDES.core
#> [1] "0.1.0"
#> 
#> $caribouMovement$version$SpaDES.tools
#> [1] "0.1.0"
#> 
#> $caribouMovement$version$caribouMovement
#> [1] "1.6.0"
#> 
#> 
#> $caribouMovement$spatialExtent
#> class       : Extent 
#> xmin        : NA 
#> xmax        : NA 
#> ymin        : NA 
#> ymax        : NA 
#> 
#> $caribouMovement$timeframe
#> [1] NA NA
#> 
#> $caribouMovement$timeunit
#> [1] "month"
#> 
#> $caribouMovement$citation
#> list()
#> 
#> $caribouMovement$documentation
#> list()
#> 
#> $caribouMovement$reqdPkgs
#> $caribouMovement$reqdPkgs[[1]]
#> [1] "grid"
#> 
#> $caribouMovement$reqdPkgs[[2]]
#> [1] "raster"
#> 
#> $caribouMovement$reqdPkgs[[3]]
#> [1] "sp"
#> 
#> $caribouMovement$reqdPkgs[[4]]
#> [1] "stats"
#> 
#> $caribouMovement$reqdPkgs[[5]]
#> [1] "SpaDES.tools"
#> 
#> 
#> $caribouMovement$parameters
#>          paramName paramClass      default          min      max
#> 1        stackName  character    landscape           NA       NA
#> 2  moveInitialTime    numeric start(si.... start(si.... end(sim)
#> 3     moveInterval    numeric            1            1        1
#> 4                N    numeric          100           10     1000
#> 5            torus    logical        FALSE        FALSE     TRUE
#> 6 .plotInitialTime    numeric   start(sim)         -Inf      Inf
#> 7    .plotInterval    numeric            1         -Inf      Inf
#> 8 .saveInitialTime    numeric           NA         -Inf      Inf
#> 9    .saveInterval    numeric           NA         -Inf      Inf
#>                                  paramDesc
#> 1                  name of the RasterStack
#> 2    time to schedule first movement event
#> 3    time interval between movoment events
#> 4                initial number of caribou
#> 5 should the map wrap around like a torus?
#> 6        time to schedule first plot event
#> 7        time interval between plot events
#> 8        time to schedule first save event
#> 9        time interval between save events
#> 
#> $caribouMovement$inputObjects
#>                            objectName objectClass                         desc
#> 1 P(sim, "caribouMovement")$stackName RasterStack layername = "habitatQuality"
#>   sourceURL
#> 1      <NA>
#> 
#> $caribouMovement$outputObjects
#>                            objectName            objectClass
#> 1 P(sim, "caribouMovement")$stackName            RasterStack
#> 2                             caribou SpatialPointsDataFrame
#>                           desc
#> 1 layername = "habitatQuality"
#> 2                         <NA>
#> 
#> 
#> $fireSpread
#> $fireSpread$name
#> [1] "fireSpread"
#> 
#> $fireSpread$description
#> [1] "Simulate fire ignition and spread on a landscape, where spread probability varies according to percent pine. Fire size statistics are collected immediately after each burn event. Requires a global simulation parameter `stackName` be set."
#> 
#> $fireSpread$keywords
#> [1] "fire"              "percolation model" "spread algorithm" 
#> 
#> $fireSpread$childModules
#> character(0)
#> 
#> $fireSpread$authors
#> [1] "Alex M Chubaty <alexander.chubaty@canada.ca> [aut, cre]" 
#> [2] "Eliot J B McIntire <eliot.mcintire@canada.ca> [aut, cre]"
#> [3] "Steve Cumming <Steve.Cumming@sbf.ulaval.ca> [aut]"       
#> 
#> $fireSpread$version
#> $fireSpread$version$SpaDES.core
#> [1] "0.1.0"
#> 
#> $fireSpread$version$SpaDES.tools
#> [1] "0.1.0"
#> 
#> $fireSpread$version$fireSpread
#> [1] "1.6.0"
#> 
#> 
#> $fireSpread$spatialExtent
#> class       : Extent 
#> xmin        : NA 
#> xmax        : NA 
#> ymin        : NA 
#> ymax        : NA 
#> 
#> $fireSpread$timeframe
#> [1] NA NA
#> 
#> $fireSpread$timeunit
#> [1] "year"
#> 
#> $fireSpread$citation
#> list()
#> 
#> $fireSpread$documentation
#> list()
#> 
#> $fireSpread$reqdPkgs
#> $fireSpread$reqdPkgs[[1]]
#> [1] "methods"
#> 
#> $fireSpread$reqdPkgs[[2]]
#> [1] "raster"
#> 
#> $fireSpread$reqdPkgs[[3]]
#> [1] "RColorBrewer"
#> 
#> $fireSpread$reqdPkgs[[4]]
#> [1] "SpaDES.tools"
#> 
#> 
#> $fireSpread$parameters
#>           paramName paramClass      default        min          max
#> 1         stackName  character    landscape         NA           NA
#> 2            nFires    numeric           10          1          100
#> 3               its    numeric        1e+06      1e+06        1e+06
#> 4       persistprob    numeric            0          0            1
#> 5    returnInterval    numeric            1          1            1
#> 6        spreadprob    numeric        0.225       0.05          0.5
#> 7         startTime    numeric start(si....          0     end(sim)
#> 8  .plotInitialTime    numeric   start(sim) start(sim) end(sim) + 1
#> 9     .plotInterval    numeric            1          1            1
#> 10 .saveInitialTime    numeric           NA         NA           NA
#> 11    .saveInterval    numeric           NA         NA           NA
#>                                     paramDesc
#> 1                     name of the RasterStack
#> 2                 number of fires to initiate
#> 3        number of iterations for fire spread
#> 4   probability of fire persisting in a pixel
#> 5                        fire return interval
#> 6  probability of fire spreading into a pixel
#> 7               time of initial fire ignition
#> 8           time to schedule first plot event
#> 9           time interval between plot events
#> 10          time to schedule first save event
#> 11          time interval between save events
#> 
#> $fireSpread$inputObjects
#>                       objectName objectClass desc sourceURL
#> 1 P(sim, "fireSpread")$stackName RasterStack <NA>      <NA>
#> 2         globals(sim)$burnStats     numeric <NA>      <NA>
#> 
#> $fireSpread$outputObjects
#>                       objectName objectClass desc other
#> 1 P(sim, "fireSpread")$stackName RasterStack <NA>    NA
#> 2         globals(sim)$burnStats     numeric <NA>    NA
#> 
#>