Parse and extract a module's version

Parse and extract a module's version

moduleVersion(module, path, sim, envir = NULL)

# S4 method for character,character,missing
moduleVersion(module, path, envir)

# S4 method for character,missing,missing
moduleVersion(module, envir)

# S4 method for character,missing,simList
moduleVersion(module, sim, envir)

Arguments

module
Character string. Your module's name.

path
Character string specifying the file path to modules directory. Default is to use the spades.modulePath option.

sim
A simList simulation object, generally produced by simInit.

envir
Optional environment in which to store parsed code. This may be useful if the same file is being parsed multiple times. This function will check in that envir for the parsed file before parsing again. If the envir is transient, then this will have no effect.

Value

numeric_version indicating the module's version.

See also

Examples

path <- system.file("sampleModules", package = "SpaDES.core")

# using filepath
moduleVersion("caribouMovement", path)

#> [1] ‘1.6.0’

# using simList
mySim <- simInit(
   times = list(start = 2000.0, end = 2002.0, timeunit = "year"),
   params = list(
     .globals = list(stackName = "landscape", burnStats = "nPixelsBurned")
   ),
   modules = list("caribouMovement"),
   paths = list(modulePath = path)
)

#> Setting:
#>   options(
#>     spades.modulePath = '/tmp/RtmplnBpLx/temp_libpath168b310a8809c/SpaDES.core/sampleModules'
#>   )

#> Paths set to:
#>   options(
#>     rasterTmpDir = '/tmp/RtmpsU1PUh/raster'
#>     reproducible.cachePath = '/tmp/RtmpsU1PUh/reproducible/cache'
#>     spades.inputPath = '/tmp/Rtmp9AsuiQ/SpaDES/inputs'
#>     spades.outputPath = '/tmp/Rtmp9AsuiQ/SpaDES/outputs'
#>     spades.modulePath = '/tmp/RtmplnBpLx/temp_libpath168b310a8809c/SpaDES.core/sampleModules'
#>   )

#> caribouMovement: module code: landscape is declared in metadata inputObjects, but no default(s) is provided in .inputObjects

#> caribouMovement: inputObjects: habitatQuality is used from sim inside Move, but is not declared in metadata inputObjects

#> Global parameter(s) not used in any module: burnStats.

moduleVersion("caribouMovement", sim = mySim)

#> [1] ‘1.6.0’