Simulation object dependency diagram

Create a sequence diagram illustrating the data object dependencies of a simulation. Offers a more detailed view of specific objects than does plotting the depsEdgeList directly with moduleDiagram().

objectDiagram(sim, ...)

# S4 method for simList
objectDiagram(sim, ...)

Arguments

sim

A simList object (typically corresponding to a completed simulation).

...

Additional arguments passed to DiagrammeR::mermaid. Useful for specifying height and width.

Value

Plots a sequence diagram, invisibly returning a DiagrammeR::mermaid object.

See also

Author

Alex Chubaty

Examples

# \donttest{
if (requireNamespace("DiagrammeR", quietly = TRUE)) {
  sim <- simInit()
  objectDiagram(sim)
  # if there are lots of objects, may need to increase width and/or height
  objectDiagram(sim, height = 3000, width = 3000)
}
#> Setting:
#>   options(
#>     reproducible.cachePath = '/tmp/RtmpchrTJY/myProject/cache'
#>     spades.inputPath = '/tmp/RtmpchrTJY/myProject/inputs'
#>     spades.outputPath = '/tmp/RtmpchrTJY/myProject/outputs'
#>     spades.modulePath = '/tmp/RtmpchrTJY/myProject/modules'
#>     spades.scratchPath = '/tmp/RtmpchrTJY/SpaDES/scratch'
#>   )
#> Nov09 16:22:33 simInit Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> Elpsed time for simInit: 0.03993511 secs

# }