Get and set simulation parameters.

params and P access the parameter slot in the simList. params has a replace method, so can be used to update a parameter value. P is a concise way to access parameters within a module. It works more like a namespaced function in the sense that the module from which it is called is the default place it will look for the parameter. To access a parameter from within a module, you can use P(sim)$paramName instead of params(sim)$moduleName$paramName

params(sim)

# S4 method for .simList
params(sim)

params(sim) <- value

# S4 method for .simList
params(sim) <- value

P(sim, module, param)

parameters(sim, asDF = FALSE)

# S4 method for .simList
parameters(sim, asDF = FALSE)

Arguments

sim
A simList object from which to extract element(s) or in which to replace element(s).

value
The object to be stored at the slot.

module
Optional character string indicating which module params should come from.

param
Optional character string indicating which parameter to choose.

asDF
Logical. For parameters, if TRUE, this will produce a single data.frame of all model parameters. If FALSE, then it will return a data.frame with 1 row for each parameter within nested lists, with the same structure as params.

Value

Returns or sets the value of the slot from the simList object.

Note

The differences between P, params and being explicit with passing arguments are mostly a question of speed and code compactness. The computationally fastest way to get a parameter is to specify moduleName and parameter name, as in: P(sim, "moduleName", "paramName") (replacing moduleName and paramName with your specific module and parameter names), but it is more verbose than P(sim)$paramName. Note: the important part for speed (e.g., 2-4x faster) is specifying the moduleName. Specifying the parameter name is <5

See also

SpaDES.core-package, specifically the section 1.2.1 on Simulation parameters.

Other functions to access elements of a simList object: .addDepends, doEvent.checkpoint, envir, events, globals, inputs, ls.simList, ls.str.simList, modules, objs, packages, paths, progressInterval, times

Examples

# NOT RUN {
modules <- list("randomLandscapes")
paths <- list(modulePath = system.file("sampleModules", package = "SpaDES.core"))
mySim <- simInit(modules = modules, paths = paths,
                 params = list(.globals = list(stackName = "landscape")))
parameters(mySim)

# }