Get and set simulation progress bar details

The progress bar can be set in two ways in SpaDES. First, by setting values in the .progress list element in the params list element passed to simInit(). Second, at the spades() call itself, which can be simpler. See examples.

progressInterval(sim)

# S4 method for class 'simList'
progressInterval(sim)

progressInterval(sim) <- value

# S4 method for class 'simList'
progressInterval(sim) <- value

progressType(sim)

# S4 method for class 'simList'
progressType(sim)

progressType(sim) <- value

# S4 method for class 'simList'
progressType(sim) <- value

Arguments

sim

A simList object from which to extract element(s) or in which to replace element(s).

value

The parameter value to be set (in the corresponding module and param).

Value

for progressInterval, a numeric corresponding to the progress update interval; for progressInterval<-, an updated simList object.

Details

Progress Bar: Progress type can be one of "text", "graphical", or "shiny". Progress interval can be a numeric. These both can get set by passing a .progress = list(type = "graphical", interval = 1) into the simInit call. See examples.

See also

Examples

# \donttest{
if (requireNamespace("SpaDES.tools", quietly = TRUE) &&
requireNamespace("NLMR", quietly = TRUE)) {
opts <- options("spades.moduleCodeChecks" = FALSE) # not necessary for example
mySim <- simInit(
  times = list(start=0.0, end=100.0),
  params = list(.globals = list(stackName = "landscape"),
  .progress = list(type = "text", interval = 10),
  checkpoint = list(interval = 10, file = "chkpnt.RData")),
  modules = list("randomLandscapes"),
  paths = list(modulePath = getSampleModules(tempdir()))
)

# progress bar
progressType(mySim) # "text"
progressInterval(mySim) # 10

# parameters
params(mySim) # returns all parameters in all modules
              # including .global, .progress, checkpoint
globals(mySim) # returns only global parameters

# checkpoint
checkpointFile(mySim) # returns the name of the checkpoint file
                      # In this example, "chkpnt.RData"
checkpointInterval(mySim) # 10

options(opts) # reset
}# }
#> Setting:
#>   options(
#>     spades.modulePath = '/tmp/RtmpEnR9g1/sampleModules'
#>   )
#> Paths set to:
#>   options(
#>     rasterTmpDir = '/tmp/RtmpEnR9g1/SpaDES/scratch/raster'
#>     reproducible.cachePath = '/tmp/RtmpEnR9g1/myProject/cache'
#>     spades.inputPath = '/tmp/RtmpEnR9g1/myProject/inputs'
#>     spades.outputPath = '/tmp/RtmpEnR9g1/myProject/outputs'
#>     spades.modulePath = '/tmp/RtmpEnR9g1/sampleModules'
#>     spades.scratchPath = '/tmp/RtmpEnR9g1/SpaDES/scratch'
#>   )
#>   terra::terraOptions(tempdir = '/tmp/RtmpEnR9g1/SpaDES/scratch/terra'
#> Sep02 20:57:34 simInit Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> Sep02 20:57:34 simInit The following .globals were used:
#> Sep02 20:57:34 simInit Key: <global, module>
#> Sep02 20:57:34 simInit               module    global
#> Sep02 20:57:34 simInit               <char>    <char>
#> Sep02 20:57:34 simInit  1: randomLandscapes stackName
#> Elapsed time for simInit: 0.05154443 secs