Get and set default working directories

Wrapper functions to access the packages options for default working directories. Note: there is an active binding made to Paths, so a user can use Paths$cachePath for example instead of getPaths()$cachePath

.paths()

getPaths()

Paths

setPaths(cachePath, inputPath, modulePath, outputPath, silent = FALSE)

Arguments

cachePath
The default local directory in which to cache simulation outputs. If not specified, defaults to getOption("reproducible.cachePath").

inputPath
The default local directory in which to look for simulation inputs If not specified, defaults to getOption("spades.inputPath").

modulePath
The default local directory where modules and data will be downloaded and stored. If not specified, defaults to getOption("spades.modulePath").

outputPath
The default local directory in which to save simulation outputs. If not specified, defaults to getOption("spades.outputPath").

silent
Logical. Should the messaging occur.

Format

An object of class list of length 4.

Value

Returns a named list of the user's default working directories. setPaths is invoked for the side effect of setting these directories.

Examples

# NOT RUN {
getPaths()                       ## returns the current default working paths
setPaths(cachePath = tempdir())  ## sets custom cachePath with other paths default
setPaths(inputPath = tempdir())  ## sets custom inputPath with other paths default
setPaths(modulePath = tempdir()) ## sets custom modulePath with other paths default
setPaths(outputPath = tempdir()) ## sets custom outputPath with other paths default

# NOTE: on loading and attaching SpaDES.core,
# an active binding is made to "Paths"

getPaths()
Paths # same
setPaths(outputPath = tempdir())
Paths # shows change
# }