Simulation environment

Accessor functions for the .xData slot, which is the default virtual slot for an S4 class object that inherits from an S3 object (specifically, the simList inherits from environment) in a simList object. These are included for advanced users.

envir(sim)

# S4 method for simList
envir(sim)

envir(sim) <- value

# S4 method for simList
envir(sim) <- value

Arguments

sim
A simList object from which to extract element(s) or in which to replace element(s).

value
The object to be stored at the slot.

Value

Returns or sets the value of the slot from the simList object.

Details

Currently, only get and set methods are defined. Subset methods are not.

See also