Metadata accessors

These accessors extract the metadata for a module (if specified) or all modules in a simList if not specified.

inputObjects(sim, module, path)

# S4 method for simList
inputObjects(sim, module, path)

# S4 method for missing
inputObjects(sim, module, path)

outputObjects(sim, module, path)

# S4 method for simList
outputObjects(sim, module, path)

# S4 method for missing
outputObjects(sim, module, path)

outputObjectNames(sim, module)

# S4 method for simList
outputObjectNames(sim, module)

reqdPkgs(sim, module, modulePath)

# S4 method for simList
reqdPkgs(sim, module, modulePath)

# S4 method for missing
reqdPkgs(sim, module, modulePath)

documentation(sim, module)

# S4 method for simList
documentation(sim, module)

citation(package, lib.loc = NULL, auto = NULL, module = character())

# S4 method for simList
citation(package, lib.loc = NULL, auto = NULL, module = character())

# S4 method for character
citation(package, lib.loc = NULL, auto = NULL, module = character())

Arguments

sim

A simList object from which to extract element(s) or in which to replace element(s).

module

Character vector of module name(s)

path

The path to the module., i.e., the modulePath. Only relevant if sim not supplied.

modulePath

That path where module can be found. If set already using setPaths, it will use that. This will be ignored if sim is supplied and is required if sim not supplied

package

For compatibility with citation. This can be a simList or a character string for a package name.

lib.loc

a character vector with path names of R libraries, or the directory containing the source for package, or NULL. The default value of NULL corresponds to all libraries currently known. If the default is used, the loaded packages are searched before the libraries.

auto

a logical indicating whether the default citation auto-generated from the package DESCRIPTION metadata should be used or not, or NULL (default), indicating that a CITATION file is used if it exists, or an object of class "packageDescription" with package metadata (see below).

Examples

if (FALSE) {
# To pre-install and pre-load all packages prior to \code{simInit}.

# set modulePath
setPaths(modulePath = system.file("sampleModules", package = "SpaDES.core"))
# use Require and reqdPkgs
if (!interactive()) chooseCRANmirror(ind = 1) #
pkgs <- reqdPkgs(module = c("caribouMovement", "randomLandscapes", "fireSpread"))
pkgs <- unique(unlist(pkgs))
Require(pkgs)
}