Extract an intact simList but with subset of objects

This is copies the non-object components of a simList (e.g., events, etc.) then selects only the objects listed in i using Copy(mget(i, envir(sim))) and adds them to the returned simList.

# S4 method for simList,character,ANY
[(x, i, j, ..., drop = TRUE)

Arguments

x

A simList

i

A character vector of objects to select.

j

Not used.

...

Not used.

drop

Not used.

Value

The [ method returns a complete simList class with all the slots copied from the original, but only the named objects in i are returned.

Author

Eliot McIntire

Examples

s <- simInit()
#> Setting:
#>   options(
#>     reproducible.cachePath = '/tmp/Rtmp7PF6uB/cache'
#>     spades.inputPath = '/tmp/Rtmp7PF6uB/inputs'
#>     spades.outputPath = '/tmp/Rtmp7PF6uB'
#>     spades.modulePath = '/tmp/Rtmp7PF6uB/sampleModules'
#>     spades.scratchPath = '/tmp/Rtmp7PF6uB/scratch'
#>   )
#> Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
s$a <- 1
s$b <- 2
s$d <- 3
s[c("a", "d")] # a simList with only 2 objects
#> ================================================================================
#> >> Simulation dependencies:
#> use `depends(sim)` to view dependencies for each module
#> 
#> >> Simulation times:
#>      current start end timeunit
#> [1,] 0       0     10  "year"  
#> 
#> >> Modules:
#>      Name Timeunit
#> 
#> >> Objects Loaded:
#> 
#> >> Objects stored:
#> a :  num 1
#> d :  num 3
#> 
#> >> Parameters:
#>      Module Parameter Value
#>  checkpoint      file      
#>  checkpoint  interval    NA
#> 
#> >> Completed Events:
#> Null data.table (0 rows and 0 cols)
#> 
#> >> Current Event:
#> Null data.table (0 rows and 0 cols)
#> 
#> >> Scheduled Events:
#>    eventTime moduleName eventType eventPriority
#> 1:         0 checkpoint      init             0
#> 2:         0       save      init             0
#> 3:         0   progress      init             0
#> 4:         0       load      init             0
#>