Zip many of the files in a simList

Currently, this will save the raster-backed files, outputs(sim), inputs(sim). It will add these to a temp file, using Copy, where appropriate to not affect the original simList. VERY experimental; unlikely to work perfectly at the moment.

zipSimList(sim, zipfile, ..., outputs = TRUE, inputs = TRUE, cache = FALSE)

Arguments

sim
A simList at the core of the zipping.

zipfile
A character string indicating the filename for the zip file. Passed to zip.

...
passed to saveSimList, including important ones such as filename.

outputs
Logical. If TRUE, all files identified in outputs(sim) will be included in the zip.

inputs
Logical. If TRUE, all files identified in inputs(sim) will be included in the zip.

cache
Logical. If TRUE, all files in cachePath(sim) will be included in the zip archive. Defaults to FALSE as this could be large, and may include many out of date elements. See Details. Not yet implemented.

Details

If cache is used, it is likely that it should be trimmed before zipping, to include only cache elements that are relevant.