Parse and extract module metadata

Parse and extract module metadata

moduleMetadata(module, path, sim)

# S4 method for character,character,missing
moduleMetadata(module, path)

# S4 method for character,missing,missing
moduleMetadata(module)

# S4 method for ANY,missing,simList
moduleMetadata(module, sim)

Arguments

module
Character string. Your module's name.

path
Character string specifying the file path to modules directory. Default is to use the spades.modulePath option.

sim
A simList simulation object, generally produced by simInit.

Value

A list of module metadata, matching the structure in defineModule.

See also

defineModule

Examples

path <- system.file("sampleModules", package = "SpaDES.core")
sampleModules <- dir(path)
x <- moduleMetadata(sampleModules[3], path)

# using simList
mySim <- simInit(
   times = list(start = 2000.0, end = 2002.0, timeunit = "year"),
   params = list(
     .globals = list(stackName = "landscape")
   ),
   modules = list("fireSpread"),
   paths = list(modulePath = system.file("sampleModules", package = "SpaDES.core"))
)
moduleMetadata(sim = mySim)

#> $name
#> [1] "fireSpread"
#> 
#> $description
#> [1] "Simulate fire ignition and spread on a landscape, where spread probability varies according to percent pine. Fire size statistics are collected immediately after each burn event. Requires a global simulation parameter `stackName` be set."
#> 
#> $keywords
#> [1] "fire"              "percolation model" "spread algorithm" 
#> 
#> $childModules
#> character(0)
#> 
#> $authors
#> [1] "Alex M Chubaty <alexander.chubaty@canada.ca> [aut, cre]" 
#> [2] "Eliot J B McIntire <eliot.mcintire@canada.ca> [aut, cre]"
#> [3] "Steve Cumming <Steve.Cumming@sbf.ulaval.ca> [aut]"       
#> 
#> $version
#> $version$SpaDES.core
#> [1] "0.1.0"
#> 
#> $version$SpaDES.tools
#> [1] "0.1.0"
#> 
#> $version$fireSpread
#> [1] "1.6.0"
#> 
#> 
#> $spatialExtent
#> class       : Extent 
#> xmin        : NA 
#> xmax        : NA 
#> ymin        : NA 
#> ymax        : NA 
#> 
#> $timeframe
#> [1] NA NA
#> 
#> $timeunit
#> [1] "year"
#> 
#> $citation
#> list()
#> 
#> $documentation
#> list()
#> 
#> $reqdPkgs
#> $reqdPkgs[[1]]
#> [1] "methods"
#> 
#> $reqdPkgs[[2]]
#> [1] "raster"
#> 
#> $reqdPkgs[[3]]
#> [1] "RColorBrewer"
#> 
#> $reqdPkgs[[4]]
#> [1] "SpaDES.tools"
#> 
#> 
#> $parameters
#>           paramName paramClass      default        min          max
#> 1         stackName  character    landscape         NA           NA
#> 2            nFires    numeric           10          1          100
#> 3               its    numeric        1e+06      1e+06        1e+06
#> 4       persistprob    numeric            0          0            1
#> 5    returnInterval    numeric            1          1            1
#> 6        spreadprob    numeric        0.225       0.05          0.5
#> 7         startTime    numeric start(si....          0     end(sim)
#> 8  .plotInitialTime    numeric   start(sim) start(sim) end(sim) + 1
#> 9     .plotInterval    numeric            1          1            1
#> 10 .saveInitialTime    numeric           NA         NA           NA
#> 11    .saveInterval    numeric           NA         NA           NA
#>                                     paramDesc
#> 1                     name of the RasterStack
#> 2                 number of fires to initiate
#> 3        number of iterations for fire spread
#> 4   probability of fire persisting in a pixel
#> 5                        fire return interval
#> 6  probability of fire spreading into a pixel
#> 7               time of initial fire ignition
#> 8           time to schedule first plot event
#> 9           time interval between plot events
#> 10          time to schedule first save event
#> 11          time interval between save events
#> 
#> $inputObjects
#>                       objectName objectClass desc sourceURL
#> 1 P(sim, "fireSpread")$stackName RasterStack <NA>      <NA>
#> 2         globals(sim)$burnStats     numeric <NA>      <NA>
#> 
#> $outputObjects
#>                       objectName objectClass desc other
#> 1 P(sim, "fireSpread")$stackName RasterStack <NA>    NA
#> 2         globals(sim)$burnStats     numeric <NA>    NA
#>