Parse and initialize a module

Internal function, used during simInit.

.parseModulePartial(sim, modules, filename, defineModuleElement, envir = NULL)

# S4 method for missing,missing,character,character
.parseModulePartial(filename, defineModuleElement, envir)

# S4 method for simList,list,missing,character
.parseModulePartial(sim, modules, defineModuleElement, envir = NULL)

.parseModule(
  sim,
  modules,
  userSuppliedObjNames = NULL,
  envir = NULL,
  notOlderThan,
  ...
)

# S4 method for simList,list
.parseModule(
  sim,
  modules,
  userSuppliedObjNames = NULL,
  envir = NULL,
  notOlderThan,
  ...
)

Arguments

sim
A simList simulation object.

modules
A list of modules with a logical attribute "parsed".

filename
The filename of the module to be parsed.

defineModuleElement
Character string indicating which of the list elements in defineModule should be extracted

envir
Optional environment in which to store parsed code. This may be useful if the same file is being parsed multiple times. This function will check in that envir for the parsed file before parsing again. If the envir is transient, then this will have no effect.

userSuppliedObjNames
Character string (or NULL, the default) indicating the names of objects that user has passed into simInit via objects or inputs. If all module inputObject dependencies are provided by user, then the .inputObjects code will be skipped.

notOlderThan
Passed to Cache that may be used for .inputObjects function call.

...
All simInit parameters.

Value

.parseModulePartial extracts just the individual element requested from the module. This can be useful if parsing the whole module would cause an error.

A simList simulation object.