.robustDigest for simList class objects

This is intended to be used within the Cache function, but can be used to evaluate what a simList would look like once it is converted to a repeatably digestible object.

# S4 method for simList
.robustDigest(object, objects,
  compareRasterFileLength = 1e+06, algo = "xxhash64",
  digestPathContent = FALSE, classOptions = list())

Arguments

object
an object to digest.

objects
Optional character vector indicating which objects are to be considered while making digestible. This is only relevant if the object being passed is an environment or list or the like.

compareRasterFileLength
Numeric. Optional. When there are Rasters, that have file-backed storage, this is passed to the length arg in digest when determining if the Raster file is already in the database. Note: uses digest for file-backed Raster. Default 1e6. Passed to .prepareFileBackedRaster.

algo
The algorithms to be used; currently available choices are md5, which is also the default, sha1, crc32, sha256, sha512, xxhash32, xxhash64 and murmur32.

digestPathContent
Logical. Should arguments that are of class Path (see examples below) have their name digested (FALSE; default), or their file contents (TRUE).

classOptions
Optional list. This will pass into .robustDigest for specific classes. Should be options that the .robustDigest knows what to do with.

Details

See .robustDigest. This method strips out stuff from a simList class object that would make it otherwise not reproducibly digestible between sessions, operating systems, or machines. This will likely still not allow identical digest results across R versions.

See also

.robustDigest