Save objects using .saveObjects in params slot of simInit

In the simInit() call, a parameter called .saveObjects can be provided in each module. This must be a character string vector of all object names to save. These objects will then be saved whenever a call to saveFiles is made.

saveFiles(sim)

Arguments

sim

A simList simulation object.

Value

(invisibly) the modified sim object. invoked for side effect of saving the simulation to file.

Details

The file names will be equal to the object name plus time(sim) is appended at the end. The files are saved as .rds files, meaning, only one object gets saved per file.

For objects saved using this function, the module developer must create save events that schedule a call to saveFiles.

If this function is used outside of a module, it will save all files in the outputs(sim) that are scheduled to be saved at the current time in the simList.

There are several ways to save objects using SpaDES.

Note

It is not possible to schedule separate saving events for each object that is listed in the .saveObjects.

Model-level saving

Using the outputs slot in the simInit() call. See example in simInit(). This can be convenient because it gives overall control of many modules at a time, and it gets automatically scheduled during the simInit() call.

Module-level saving

Using the saveFiles function inside a module. This must be accompanied by a .saveObjects vector or list element in the params slot in the simList(). Usually a module developer will create this method for future users of their module.

Custom saving

A module developer can save any object at any time inside their module, using standard R functions for saving R objects (e.g., save or saveRDS). This is the least modular approach, as it will happen whether a module user wants it or not.

Author

Eliot McIntire and Alex Chubaty

Examples

# \donttest{
if (requireNamespace("SpaDES.tools", quietly = TRUE) &&
requireNamespace("NLMR", quietly = TRUE)) {
## This will save the "caribou" object at the save interval of 1 unit of time
## in the outputPath location
  outputPath <- file.path(tempdir(), "test_save")
  times <- list(start = 0, end = 1, "month")
  parameters <- list(
    .globals = list(stackName = "landscape"),
    caribouMovement = list(
      .saveObjects = "caribou",
      .saveInitialTime = 1, .saveInterval = 1,
      .plots = NA
    ),
    randomLandscapes = list(.plots = NA, nx = 20, ny = 20))

  modules <- list("randomLandscapes", "caribouMovement")
  paths <- list(
    modulePath = getSampleModules(tempdir()),
    outputPath = outputPath
  )
  opts <- options("spades.moduleCodeChecks" = FALSE) # not necessary for example
  mySim <- simInit(times = times, params = parameters, modules = modules,
                   paths = paths)

  # The caribou module has a saveFiles(sim) call, so it will save caribou
  spades(mySim)
  dir(outputPath)

  # remove the files
  file.remove(dir(outputPath, full.names = TRUE))

  ## save multiple outputs
  parameters <- list(
    .globals = list(stackName = "landscape"),
    caribouMovement = list(
      .saveObjects = c("caribou", "habitatQuality"),
      .saveInitialTime = 1, .saveInterval = 1,
      .plots = NA
    ),
    randomLandscapes = list(.plots = NA, nx = 20, ny = 20))

  mySim <- simInit(times = times, params = parameters, modules = modules,
                 paths = paths)

  spades(mySim)
  dir(outputPath)
  # remove the files
  file.remove(dir(outputPath, full.names = TRUE))

  options(opts) # clean up

}# }
#> Setting:
#>   options(
#>     spades.outputPath = '/tmp/RtmpchrTJY/test_save'
#>     spades.modulePath = '/tmp/RtmpchrTJY/sampleModules'
#>   )
#> Paths set to:
#>   options(
#>     rasterTmpDir = '/tmp/RtmpchrTJY/SpaDES/scratch/raster'
#>     reproducible.cachePath = '/tmp/RtmpchrTJY/myProject/cache'
#>     spades.inputPath = '/tmp/RtmpchrTJY/myProject/inputs'
#>     spades.outputPath = '/tmp/RtmpchrTJY/test_save'
#>     spades.modulePath = '/tmp/RtmpchrTJY/sampleModules'
#>     spades.scratchPath = '/tmp/RtmpchrTJY/SpaDES/scratch'
#>   )
#>   terra::terraOptions(tempdir = '/tmp/RtmpchrTJY/SpaDES/scratch/terra'
#> Nov09 16:22:41 simInit Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> Nov09 16:22:41 simInit The following .globals were used:
#> Nov09 16:22:41 simInit              module    global
#> Nov09 16:22:41 simInit 1:  caribouMovement stackName
#> Nov09 16:22:41 simInit 2: randomLandscapes stackName
#> Elpsed time for simInit: 0.1221495 secs
#> Nov09 16:22:41 simInit Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> Nov09 16:22:41 chckpn:init total elpsd: 0.12 secs | 0 checkpoint init 0
#> Nov09 16:22:41 save  :init total elpsd: 0.13 secs | 0 save init 0
#> Nov09 16:22:41 prgrss:init total elpsd: 0.13 secs | 0 progress init 0
#> Nov09 16:22:41 load  :init total elpsd: 0.13 secs | 0 load init 0
#> Nov09 16:22:41 rndmLn:init total elpsd: 0.14 secs | 0 randomLandscapes init 1
#> Nov09 16:22:42 rndmLn:init New objects created:
#> Nov09 16:22:42 rndmLn:init 1:            landscape
#> Nov09 16:22:42 crbMvm:init total elpsd: 0.82 secs | 0 caribouMovement init 1
#> Nov09 16:22:42 crbMvm:init New objects created:
#> Nov09 16:22:42 crbMvm:init 1:              caribou
#> Nov09 16:22:42 rndmLn:plot total elpsd: 0.84 secs | 0 randomLandscapes plot 1
#> Nov09 16:22:42 crbMvm:plot.init total elpsd: 0.86 secs | 0 caribouMovement plot.in
#> Nov09 16:22:42 crbMvm:move total elpsd: 0.88 secs | 1 caribouMovement move 5
#> Nov09 16:22:42 crbMvm:move New objects created:
#> Nov09 16:22:42 crbMvm:move 1:       habitatQuality
#> Nov09 16:22:42 crbMvm:plot total elpsd: 0.93 secs | 1 caribouMovement plot 10
#> Nov09 16:22:42 crbMvm:save total elpsd: 0.94 secs | 1 caribouMovement save 11
#> simList saved in
#> SpaDES.core:::.pkgEnv$.sim
#> It will be deleted at next spades() call.
#> Setting:
#>   options(
#>     spades.outputPath = '/tmp/RtmpchrTJY/test_save'
#>     spades.modulePath = '/tmp/RtmpchrTJY/sampleModules'
#>   )
#> Paths set to:
#>   options(
#>     rasterTmpDir = '/tmp/RtmpchrTJY/SpaDES/scratch/raster'
#>     reproducible.cachePath = '/tmp/RtmpchrTJY/myProject/cache'
#>     spades.inputPath = '/tmp/RtmpchrTJY/myProject/inputs'
#>     spades.outputPath = '/tmp/RtmpchrTJY/test_save'
#>     spades.modulePath = '/tmp/RtmpchrTJY/sampleModules'
#>     spades.scratchPath = '/tmp/RtmpchrTJY/SpaDES/scratch'
#>   )
#>   terra::terraOptions(tempdir = '/tmp/RtmpchrTJY/SpaDES/scratch/terra'
#> Nov09 16:22:44 simInit Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> Nov09 16:22:44 simInit The following .globals were used:
#> Nov09 16:22:44 simInit              module    global
#> Nov09 16:22:44 simInit 1:  caribouMovement stackName
#> Nov09 16:22:44 simInit 2: randomLandscapes stackName
#> Elpsed time for simInit: 0.1153014 secs
#> Nov09 16:22:44 simInit Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> Nov09 16:22:44 chckpn:init total elpsd: 0.12 secs | 0 checkpoint init 0
#> Nov09 16:22:44 save  :init total elpsd: 0.12 secs | 0 save init 0
#> Nov09 16:22:44 prgrss:init total elpsd: 0.12 secs | 0 progress init 0
#> Nov09 16:22:44 load  :init total elpsd: 0.13 secs | 0 load init 0
#> Nov09 16:22:44 rndmLn:init total elpsd: 0.13 secs | 0 randomLandscapes init 1
#> Nov09 16:22:44 rndmLn:init New objects created:
#> Nov09 16:22:44 rndmLn:init 1:            landscape
#> Nov09 16:22:44 crbMvm:init total elpsd: 0.29 secs | 0 caribouMovement init 1
#> Nov09 16:22:44 crbMvm:init New objects created:
#> Nov09 16:22:44 crbMvm:init 1:              caribou
#> Nov09 16:22:44 rndmLn:plot total elpsd: 0.33 secs | 0 randomLandscapes plot 1
#> Nov09 16:22:44 crbMvm:plot.init total elpsd: 0.33 secs | 0 caribouMovement plot.in
#> Nov09 16:22:44 crbMvm:move total elpsd: 0.34 secs | 1 caribouMovement move 5
#> Nov09 16:22:44 crbMvm:move New objects created:
#> Nov09 16:22:44 crbMvm:move 1:       habitatQuality
#> Nov09 16:22:44 crbMvm:plot total elpsd: 0.37 secs | 1 caribouMovement plot 10
#> Nov09 16:22:44 crbMvm:save total elpsd: 0.37 secs | 1 caribouMovement save 11
#> simList saved in
#> SpaDES.core:::.pkgEnv$.sim
#> It will be deleted at next spades() call.