Metadata accessors

These accessors extract the metadata for a module (if specified) or all modules in a simList if not specified.

inputObjects(sim, module)

# S4 method for simList
inputObjects(sim, module)

outputObjects(sim, module)

# S4 method for simList
outputObjects(sim, module)

outputObjectNames(sim, module)

# S4 method for simList
outputObjectNames(sim, module)

reqdPkgs(sim, module)

# S4 method for simList
reqdPkgs(sim, module)

documentation(sim, module)

# S4 method for simList
documentation(sim, module)

citation(package, lib.loc = NULL, auto = NULL, module = character())

# S4 method for simList
citation(package, lib.loc = NULL, auto = NULL, module = character())

# S4 method for character
citation(package, lib.loc = NULL, auto = NULL, module = character())

Arguments

sim
A simList object from which to extract element(s) or in which to replace element(s).

module
Optional character string indicating which module params should come from.

package
For compatibility with citation. This can be a simList or a character string for a package name.

lib.loc
a character vector with path names of R libraries, or the directory containing the source for package, or NULL. The default value of NULL corresponds to all libraries currently known. If the default is used, the loaded packages are searched before the libraries.

auto
a logical indicating whether the default citation auto-generated from the package DESCRIPTION metadata should be used or not, or NULL (default), indicating that a CITATION file is used if it exists, or an object of class "packageDescription" with package metadata (see below).