Metadata accessors

These accessors extract the metadata for a module (if specified) or all modules in a simList if not specified.

inputObjects(sim, module, path)

# S4 method for simList
inputObjects(sim, module, path)

# S4 method for missing
inputObjects(sim, module, path)

outputObjects(sim, module, path)

# S4 method for simList
outputObjects(sim, module, path)

# S4 method for missing
outputObjects(sim, module, path)

outputObjectNames(sim, module)

# S4 method for simList
outputObjectNames(sim, module)

reqdPkgs(sim, module, modulePath)

# S4 method for simList
reqdPkgs(sim, module, modulePath)

# S4 method for missing
reqdPkgs(sim, module, modulePath)

documentation(sim, module)

# S4 method for simList
documentation(sim, module)

sessInfo(sim)

# S4 method for simList
sessInfo(sim)

Arguments

sim

A simList object from which to extract element(s) or in which to replace element(s).

module

Character vector of module name(s)

path

The path to the module., i.e., the modulePath. Only relevant if sim not supplied.

modulePath

That path where module can be found. If set already using setPaths, it will use that. This will be ignored if sim is supplied and is required if sim not supplied

Examples

# \donttest{
# set modulePath
setPaths(modulePath = getSampleModules(tempdir()))
#> Setting:
#>   options(
#>     spades.modulePath = '/tmp/RtmpwrkDz7/sampleModules'
#>   )
#> Paths set to:
#>   options(
#>     rasterTmpDir = '/tmp/RtmpwrkDz7/scratch/raster'
#>     reproducible.cachePath = '/tmp/RtmpwrkDz7/cache'
#>     spades.inputPath = '/tmp/RtmpwrkDz7/inputs'
#>     spades.outputPath = '/tmp/RtmpwrkDz7'
#>     spades.modulePath = '/tmp/RtmpwrkDz7/sampleModules'
#>     spades.scratchPath = '/tmp/RtmpwrkDz7/scratch'
#>   )
#>   terra::terraOptions(tempdir = '/tmp/RtmpwrkDz7/scratch/terra'
# use Require and reqdPkgs
pkgs <- reqdPkgs(module = c("caribouMovement", "randomLandscapes", "fireSpread"))
# }