Run a spatial discrete event simulation

Here, we implement a simulation in a more modular fashion so it's easier to add submodules to the simulation. We use S4 classes and methods, and use data.table instead of data.frame to implement the event queue (because it is much faster).

spades(
  sim,
  debug = getOption("spades.debug"),
  progress = NA,
  cache,
  .plotInitialTime = NULL,
  .saveInitialTime = NULL,
  notOlderThan = NULL,
  events = NULL,
  .plots = getOption("spades.plots", NULL),
  ...
)

# S4 method for simList,ANY,ANY,missing
spades(
  sim,
  debug = getOption("spades.debug"),
  progress = NA,
  cache,
  .plotInitialTime = NULL,
  .saveInitialTime = NULL,
  notOlderThan = NULL,
  events = NULL,
  .plots = getOption("spades.plots", NULL),
  ...
)

# S4 method for ANY,ANY,ANY,logical
spades(
  sim,
  debug = getOption("spades.debug"),
  progress = NA,
  cache,
  .plotInitialTime = NULL,
  .saveInitialTime = NULL,
  notOlderThan = NULL,
  events = NULL,
  .plots = getOption("spades.plots", NULL),
  ...
)

Arguments

sim

A simList simulation object, generally produced by simInit.

debug

Optional tools for invoking debugging. Supplying a list will invoke the more powerful logging package. See details. Default is to use the value in getOption("spades.debug").

progress

Logical (TRUE or FALSE show a graphical progress bar), character ("graphical", "text") or numeric indicating the number of update intervals to show in a graphical progress bar.

cache

Logical. If TRUE, then the spades call will be cached. This means that if the call is made again with the same simList, then spades will return the return value from the previous run of that exact same simList. Default FALSE. See Details. See also the vignette on caching for examples.

.plotInitialTime

Numeric. Temporarily override the .plotInitialTime parameter for all modules. See Details.

.saveInitialTime

Numeric. Temporarily override the .plotInitialTime parameter for all modules. See Details.

notOlderThan

Date or time. Passed to reproducible::Cache to update the cache. Default is NULL, meaning don't update the cache. If Sys.time() is provided, then it will force a recache, i.e., remove old value and replace with new value. Ignored if cache is FALSE.

events

A character vector or a named list of character vectors. If specified, the simulations will only do the events indicated here. If a named list, the names must correspond to the modules and the character vectors can be specific events within each of the named modules. With the list form, all unspecified modules will run all their events, including internal spades modules, e.g., save, that get invoked with the outputs argument in simInit. See example.

.plots

Character. Sets the parameter of this name in all modules. See Plots() for possible values. The parameter is intended to slowly take over from .plotInitialTime as a mechanism to turn on or off plotting. For backwards compatibility, if .plotInitialTime is not set in this spades call, but this .plots is used, two things will happen: setting this without "screen" will turn off all plotting; setting this with "screen" will trigger plotting for any modules that use this parameter but will have no effect on other modules. To get plotting, therefore, it may be necessary to also set .plotInitialTime = start(sim).

...

Any. Can be used to make a unique cache identity, such as "replicate = 1". This will be included in the Cache call, so will be unique and thus spades will not use a cached copy as long as anything passed in ... is unique, i.e., not cached previously.

Value

Invisibly returns the modified simList object.

Details

The is the workhorse function in the SpaDES package. It runs simulations by implementing the rules outlined in the simList.

This function gives simple access to two sets of module parameters: .plotInitialTime and with .plotInitialTime. The primary use of these arguments is to temporarily turn off plotting and saving. "Temporary" means that the simList is not changed, so it can be used again with the simList values reinstated. To turn off plotting and saving, use .plotInitialTime = NA or .saveInitialTime = NA. NOTE: if a module did not use .plotInitialTime or .saveInitialTime, then these arguments will not do anything.

Note

The debug option is primarily intended to facilitate building simulation models by the user. Will print additional outputs informing the user of updates to the values of various simList slot components. See https://github.com/PredictiveEcology/SpaDES/wiki/Debugging for details.

Caching with SpaDES

There are numerous ways in which Caching can be used within SpaDES. Please see the vignette https://spades-core.predictiveecology.org/articles/iii-cache.html for many examples. Briefly, functions, events, modules, entire spades calls or experiment calls (see https://github.com/PredictiveEcology/SpaDES.experiment) can be cached and mixtures of all of these will work. For functions, simply wrap the call with Cache, moving the original function name into the first argument of Cache. For events or modules, set the module parameters, .useCache, e.g., simInit(..., parameters = list(myModule = list(.useCache = "init"))). This can be set to an event name, which will cache that event, or a logical, which will cache every event in that module. Event and module caching makes most sense when the event or module only runs once, such as an initialization or data preparation event/module. Caching an entire simulation is actually just a function call to simInitAndSpades, for example. So, simply writing Cache(simInitAndSpades, modules = ...) will effectively cache a whole simulation. Finally for experiments, it is just like a function call: Cache(simInitandExperiment, ...). The final way Caching can be done is in experiment or spades, by setting the cache argument.

If cache is TRUE, this allows for a seamless way to "save" results of a simulation. The user does not have to intentionally do any saving manually. Instead, upon a call to spades in which the simList is identical, the function will simply return the result that would have come if it had been rerun. Use this with caution, as it will return exactly the result from a previous run, even if there is stochasticity internally. Caching is only based on the input simList. See also the vignette on caching for examples.

debug

The most powerful way to use debug is to invoke the logging R package. To invoke this, debug must be a list with up to 3 named elements: console, file, and debug. Each of these list elements must be a list (including empty list() for defaults) with the sub-list elements here:

console	level	The level, see below, of information shown

file	append	Logical. If TRUE, the default, then log entries are appended to file, if it exists

file	A filename. Defaults to log.txt

level	The level, see below, of information shown

debug	See possible values below

level can be a number from 0 to 100 or a character string matching one of the values in logging::loglevels. These are hierarchical levels of information passed to the console. Set a lower number for more information and a higher number for less information. Errors in code will be shown if level is set to "ERROR" or 40 or above; warnings in code will be shown if level is set to "WARN" or 30 or above; normal messages in code will be shown if level is set to "INFO" or 20 or above. For consistency with base R messaging, if default level is used, then normal messaging via message will be shown; this means that suppressMessages will work to suppress messaging only when level is set to "INFO" or 20. Some functions in the SpaDES ecosystem may have information at the lower levels, but currently, there are few to none.

debug is specified as a non-list argument to spades or as list(debug = ...), then it can be a logical, a quoted call, a character vector or a numeric scalar (currently 1 or 2) or a list of any of these to get multiple outputs. This will be run at the start of every event. The following options for debug are available. Each of these can also be in a list to get multiple outputs:

TRUE	current(sim) will be printed at the start of each event as it runs

a function name (as character string)	If a function, then it will be run on the simList, e.g., "time" will run time(sim) at each event.

moduleName (as character string)	All calls to that module will be entered interactively

eventName (as character string)	All calls that have that event name (in any module) will be entered interactively

c(<moduleName>, <eventName>)	Only the event in that specified module will be entered into.

Any other R expression expressed as a character string or quoted call	Will be evaluated with access to the simList as sim. If this is more than one character string, then all will be printed to the screen in their sequence.

A numeric scalar, currently 1 or 2 (maybe others)	This will print out alternative forms of event information that users may find useful

If not specified in the function call, the package option spades.debug is used.

If options("spades.browserOnError" = TRUE) (experimental still) if there is an error, it will attempt to open a browser in the event where the error occurred. You can edit, and then press c to continue or Q to quit, plus all other normal interactive browser tools. c will trigger a reparse and events will continue as scheduled, starting with the one just edited. There may be some unexpected consequences if the simList objects had already been changed before the error occurred.

References

Matloff, N. (2011). The Art of R Programming (ch. 7.8.3). San Francisco, CA: No Starch Press, Inc.. Retrieved from https://nostarch.com/artofr.htm

See also

Author

Alex Chubaty and Eliot McIntire

Examples

# \donttest{
if (requireNamespace("SpaDES.tools", quietly = TRUE) &&
    requireNamespace("NLMR", quietly = TRUE)) {
  # some options are not necessary when not interactive
  opts <- options("spades.moduleCodeChecks" = FALSE, "spades.useRequire" = FALSE)
  if (!interactive()) opts <- append(opts, options("spades.plots" = NA,
                                                   "spades.debug" = FALSE))
  mySim <- simInit(
   times = list(start = 0.0, end = 1.0, timeunit = "year"),
   params = list(
     randomLandscapes = list(nx = 10, ny = 10),
     .globals = list(stackName = "landscape", burnStats = "nPixelsBurned",
                     .plots = NA) # plotting off --> not relevant for example
   ),
   modules = list("randomLandscapes", "fireSpread", "caribouMovement"),
   paths = list(modulePath = system.file("sampleModules", package = "SpaDES.core"))
  )
  spades(mySim)

  # Different debug options (overrides the package option 'spades.debug')
  spades(mySim, debug = TRUE) # Fastest
  spades(mySim, debug = "print(table(sim$landscape$Fires[]))")
  # To get a combination -- use list(debug = list(..., ...))
  spades(mySim, debug = list(debug = list(1, quote(as.data.frame(table(sim$landscape$Fires[]))))))

  # Can turn off plotting at spades call, and inspect the output simList instead
  out <- spades(mySim, .plots = NA)
  completed(out) # shows completed events

  # use cache -- simInit should generally be rerun each time a spades call is made
  #   to guarantee that it is identical. Here, run spades call twice, first
  #   time to establish cache, second time to return cached result
  for (i in 1:2) {
   mySim <- simInit(
     times = list(start = 0.0, end = 1.0, timeunit = "year"),
     params = list(
       randomLandscapes = list(nx = 10, ny = 10),
       .globals = list(stackName = "landscape", burnStats = "nPixelsBurned")
     ),
     modules = list("randomLandscapes", "fireSpread", "caribouMovement"),
     paths = list(modulePath = system.file("sampleModules", package = "SpaDES.core"))
   )
   print(system.time(out <- spades(mySim, cache = TRUE, .plots = NA)))
  }

  # E.g., with only the init events
  outInitsOnly <- spades(mySim, events = "init")

  # or more fine grained control
  outSomeEvents <- spades(mySim, .plots = NA,
          events = list(randomLandscapes = c("init"),
                        fireSpread = c("init", "burn")))

  # with outputs, the save module gets invoked and must be explicitly limited to "init"
  mySim <- simInit(
   times = list(start = 0.0, end = 1.0, timeunit = "year"),
   params = list(
     randomLandscapes = list(nx = 10, ny = 10),
     .globals = list(stackName = "landscape", burnStats = "nPixelsBurned")
   ),
   modules = list("randomLandscapes", "fireSpread", "caribouMovement"),
   outputs = data.frame(objectName = "landscape", saveTime = 0:2),
   paths = list(modulePath = system.file("sampleModules", package = "SpaDES.core"))
  )
  # This will print a message saying that caribouMovement will run its events
  outSomeEvents <- spades(mySim, .plots = NA,
          events = list(randomLandscapes = c("init"),
                        fireSpread = c("init", "burn"),
                        save = "init"))

  options(opts) # reset options
}
#> Setting:
#>   options(
#>     spades.modulePath = '/home/runner/work/_temp/Library/SpaDES.core/sampleModules'
#>   )
#> Paths set to:
#>   options(
#>     rasterTmpDir = '/tmp/Rtmp8X08T2/scratch/raster'
#>     reproducible.cachePath = '/tmp/Rtmp8X08T2/cache'
#>     spades.inputPath = '/tmp/Rtmp8X08T2/inputs'
#>     spades.outputPath = '/tmp/Rtmp8X08T2'
#>     spades.modulePath = '/home/runner/work/_temp/Library/SpaDES.core/sampleModules'
#>     spades.scratchPath = '/tmp/Rtmp8X08T2/scratch'
#>   )
#>   terra::terraOptions(tempdir = '/tmp/Rtmp8X08T2/scratch/terra'
#> Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> defineParameter: '.useCache' is not of specified type 'logical'.
#> The following .globals were used:
#>              module    global
#> 1:  caribouMovement    .plots
#> 2:       fireSpread    .plots
#> 3: randomLandscapes    .plots
#> 4:  caribouMovement stackName
#> 5:       fireSpread stackName
#> 6: randomLandscapes stackName
#> Jul08 01:42:24       Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> simList saved in
#>  SpaDES.core:::.pkgEnv$.sim 
#> It will be deleted at next spades() call.
#> Jul08 01:42:25       Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> Jul08 01:42:25 chckpn eventTime moduleName eventType eventPriority
#> Jul08 01:42:25 chckpn 0         checkpoint init      0            
#> Jul08 01:42:25 save   0         save       init      0            
#> Jul08 01:42:25 prgrss 0         progress   init      0            
#> Jul08 01:42:25 load   0         load       init      0            
#> Jul08 01:42:25 rndmLn 0         randomLandscapes init      1            
#> Jul08 01:42:25 frSprd 0         fireSpread       init      1            
#> Jul08 01:42:25 crbMvm 0         caribouMovement  init      1            
#> Jul08 01:42:25 frSprd 1         fireSpread       burn      5            
#> Jul08 01:42:25 crbMvm 1         caribouMovement  move      5            
#> Jul08 01:42:25 frSprd 1         fireSpread       stats     5            
#> simList saved in
#>  SpaDES.core:::.pkgEnv$.sim 
#> It will be deleted at next spades() call.
#> Jul08 01:42:25       Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> Jul08 01:42:25 chckpn  total elpsd: 0.74 secs | 0 checkpoint init 0
#> 
#>  0  1  2  3  4  5  6  7  8  9 10 
#> 48  5  1  5  1  3 10 10  5  1 11 
#> Jul08 01:42:25 chckpn 48
#> Jul08 01:42:25 chckpn 5
#> Jul08 01:42:25 chckpn 1
#> Jul08 01:42:25 chckpn 5
#> Jul08 01:42:25 chckpn 1
#> Jul08 01:42:25 chckpn 3
#> Jul08 01:42:25 chckpn 10
#> Jul08 01:42:25 chckpn 10
#> Jul08 01:42:25 chckpn 5
#> Jul08 01:42:25 chckpn 1
#> Jul08 01:42:25 chckpn 11
#> Jul08 01:42:25 save    total elpsd: 0.76 secs | 0 save init 0
#> 
#>  0  1  2  3  4  5  6  7  8  9 10 
#> 48  5  1  5  1  3 10 10  5  1 11 
#> Jul08 01:42:25 save   48
#> Jul08 01:42:25 save   5
#> Jul08 01:42:25 save   1
#> Jul08 01:42:25 save   5
#> Jul08 01:42:25 save   1
#> Jul08 01:42:25 save   3
#> Jul08 01:42:25 save   10
#> Jul08 01:42:25 save   10
#> Jul08 01:42:25 save   5
#> Jul08 01:42:25 save   1
#> Jul08 01:42:25 save   11
#> Jul08 01:42:25 prgrss  total elpsd: 0.77 secs | 0 progress init 0
#> 
#>  0  1  2  3  4  5  6  7  8  9 10 
#> 48  5  1  5  1  3 10 10  5  1 11 
#> Jul08 01:42:25 prgrss 48
#> Jul08 01:42:25 prgrss 5
#> Jul08 01:42:25 prgrss 1
#> Jul08 01:42:25 prgrss 5
#> Jul08 01:42:25 prgrss 1
#> Jul08 01:42:25 prgrss 3
#> Jul08 01:42:25 prgrss 10
#> Jul08 01:42:25 prgrss 10
#> Jul08 01:42:25 prgrss 5
#> Jul08 01:42:25 prgrss 1
#> Jul08 01:42:25 prgrss 11
#> Jul08 01:42:25 load    total elpsd: 0.79 secs | 0 load init 0
#> 
#>  0  1  2  3  4  5  6  7  8  9 10 
#> 48  5  1  5  1  3 10 10  5  1 11 
#> Jul08 01:42:25 load   48
#> Jul08 01:42:25 load   5
#> Jul08 01:42:25 load   1
#> Jul08 01:42:25 load   5
#> Jul08 01:42:25 load   1
#> Jul08 01:42:25 load   3
#> Jul08 01:42:25 load   10
#> Jul08 01:42:25 load   10
#> Jul08 01:42:25 load   5
#> Jul08 01:42:25 load   1
#> Jul08 01:42:25 load   11
#> Jul08 01:42:25 rndmLn  total elpsd: 0.81 secs | 0 randomLandscapes init 
#> 
#>  0  1  2  3  4  5  6  7  8  9 10 
#> 48  5  1  5  1  3 10 10  5  1 11 
#> Jul08 01:42:25 rndmLn 48
#> Jul08 01:42:25 rndmLn 5
#> Jul08 01:42:25 rndmLn 1
#> Jul08 01:42:25 rndmLn 5
#> Jul08 01:42:25 rndmLn 1
#> Jul08 01:42:25 rndmLn 3
#> Jul08 01:42:25 rndmLn 10
#> Jul08 01:42:25 rndmLn 10
#> Jul08 01:42:25 rndmLn 5
#> Jul08 01:42:25 rndmLn 1
#> Jul08 01:42:25 rndmLn 11
#> Jul08 01:42:25 frSprd  total elpsd: 0.98 secs | 0 fireSpread init 1
#> Error : [subset] invalid name(s)
#> Jul08 01:42:25 frSprd Error : [subset] invalid name(s)
#> Jul08 01:42:25 crbMvm  total elpsd: 1.1 secs | 0 caribouMovement init 1
#> 
#>   0 
#> 100 
#> Jul08 01:42:25 crbMvm 100
#> Jul08 01:42:25 frSprd  total elpsd: 1.1 secs | 1 fireSpread burn 5
#> 
#>   0 
#> 100 
#> Jul08 01:42:25 frSprd 100
#> Jul08 01:42:25 crbMvm  total elpsd: 1.1 secs | 1 caribouMovement move 5
#> 
#>  0  1  2  3  4  5  6  7  8  9 10 
#> 22  6 16 28 12  2  7  1  1  3  2 
#> Jul08 01:42:25 crbMvm 22
#> Jul08 01:42:25 crbMvm 6
#> Jul08 01:42:25 crbMvm 16
#> Jul08 01:42:25 crbMvm 28
#> Jul08 01:42:25 crbMvm 12
#> Jul08 01:42:25 crbMvm 2
#> Jul08 01:42:25 crbMvm 7
#> Jul08 01:42:25 crbMvm 1
#> Jul08 01:42:25 crbMvm 1
#> Jul08 01:42:25 crbMvm 3
#> Jul08 01:42:25 crbMvm 2
#> Jul08 01:42:25 frSprd  total elpsd: 1.2 secs | 1 fireSpread stats 5
#> 
#>  0  1  2  3  4  5  6  7  8  9 10 
#> 22  6 16 28 12  2  7  1  1  3  2 
#> Jul08 01:42:26 frSprd 22
#> Jul08 01:42:26 frSprd 6
#> Jul08 01:42:26 frSprd 16
#> Jul08 01:42:26 frSprd 28
#> Jul08 01:42:26 frSprd 12
#> Jul08 01:42:26 frSprd 2
#> Jul08 01:42:26 frSprd 7
#> Jul08 01:42:26 frSprd 1
#> Jul08 01:42:26 frSprd 1
#> Jul08 01:42:26 frSprd 3
#> Jul08 01:42:26 frSprd 2
#> simList saved in
#>  SpaDES.core:::.pkgEnv$.sim 
#> It will be deleted at next spades() call.
#> Jul08 01:42:26       Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> Jul08 01:42:26 chckpn  total elpsd: 1.2 secs | 0 checkpoint init 0
#> Jul08 01:42:26 chckpn  1:    0   22
#> Jul08 01:42:26 chckpn  2:    1    6
#> Jul08 01:42:26 chckpn  3:    2   16
#> Jul08 01:42:26 chckpn  4:    3   28
#> Jul08 01:42:26 chckpn  5:    4   12
#> Jul08 01:42:26 chckpn  6:    5    2
#> Jul08 01:42:26 chckpn  7:    6    7
#> Jul08 01:42:26 chckpn  8:    7    1
#> Jul08 01:42:26 chckpn  9:    8    1
#> Jul08 01:42:26 chckpn 10:    9    3
#> Jul08 01:42:26 chckpn 11:   10    2
#> Jul08 01:42:26 save    total elpsd: 1.3 secs | 0 save init 0
#> Jul08 01:42:26 save    1:    0   22
#> Jul08 01:42:26 save    2:    1    6
#> Jul08 01:42:26 save    3:    2   16
#> Jul08 01:42:26 save    4:    3   28
#> Jul08 01:42:26 save    5:    4   12
#> Jul08 01:42:26 save    6:    5    2
#> Jul08 01:42:26 save    7:    6    7
#> Jul08 01:42:26 save    8:    7    1
#> Jul08 01:42:26 save    9:    8    1
#> Jul08 01:42:26 save   10:    9    3
#> Jul08 01:42:26 save   11:   10    2
#> Jul08 01:42:26 prgrss  total elpsd: 1.3 secs | 0 progress init 0
#> Jul08 01:42:26 prgrss  1:    0   22
#> Jul08 01:42:26 prgrss  2:    1    6
#> Jul08 01:42:26 prgrss  3:    2   16
#> Jul08 01:42:26 prgrss  4:    3   28
#> Jul08 01:42:26 prgrss  5:    4   12
#> Jul08 01:42:26 prgrss  6:    5    2
#> Jul08 01:42:26 prgrss  7:    6    7
#> Jul08 01:42:26 prgrss  8:    7    1
#> Jul08 01:42:26 prgrss  9:    8    1
#> Jul08 01:42:26 prgrss 10:    9    3
#> Jul08 01:42:26 prgrss 11:   10    2
#> Jul08 01:42:26 load    total elpsd: 1.3 secs | 0 load init 0
#> Jul08 01:42:26 load    1:    0   22
#> Jul08 01:42:26 load    2:    1    6
#> Jul08 01:42:26 load    3:    2   16
#> Jul08 01:42:26 load    4:    3   28
#> Jul08 01:42:26 load    5:    4   12
#> Jul08 01:42:26 load    6:    5    2
#> Jul08 01:42:26 load    7:    6    7
#> Jul08 01:42:26 load    8:    7    1
#> Jul08 01:42:26 load    9:    8    1
#> Jul08 01:42:26 load   10:    9    3
#> Jul08 01:42:26 load   11:   10    2
#> Jul08 01:42:26 rndmLn  total elpsd: 1.4 secs | 0 randomLandscapes init 1
#> Jul08 01:42:26 rndmLn  1:    0   22
#> Jul08 01:42:26 rndmLn  2:    1    6
#> Jul08 01:42:26 rndmLn  3:    2   16
#> Jul08 01:42:26 rndmLn  4:    3   28
#> Jul08 01:42:26 rndmLn  5:    4   12
#> Jul08 01:42:26 rndmLn  6:    5    2
#> Jul08 01:42:26 rndmLn  7:    6    7
#> Jul08 01:42:26 rndmLn  8:    7    1
#> Jul08 01:42:26 rndmLn  9:    8    1
#> Jul08 01:42:26 rndmLn 10:    9    3
#> Jul08 01:42:26 rndmLn 11:   10    2
#> Jul08 01:42:26 frSprd  total elpsd: 1.6 secs | 0 fireSpread init 1
#> Error : [subset] invalid name(s)
#> Jul08 01:42:26 frSprd Error : [subset] invalid name(s)
#> Jul08 01:42:26 crbMvm  total elpsd: 1.6 secs | 0 caribouMovement init 1
#> Jul08 01:42:26 crbMvm 1:    0  100
#> Jul08 01:42:26 frSprd  total elpsd: 1.7 secs | 1 fireSpread burn 5
#> Jul08 01:42:26 frSprd 1:    0  100
#> Jul08 01:42:26 crbMvm  total elpsd: 1.7 secs | 1 caribouMovement move 5
#> Jul08 01:42:26 crbMvm  1:    0   32
#> Jul08 01:42:26 crbMvm  2:    1    3
#> Jul08 01:42:26 crbMvm  3:    2    8
#> Jul08 01:42:26 crbMvm  4:    3    1
#> Jul08 01:42:26 crbMvm  5:    4   11
#> Jul08 01:42:26 crbMvm  6:    5    4
#> Jul08 01:42:26 crbMvm  7:    6    7
#> Jul08 01:42:26 crbMvm  8:    7   23
#> Jul08 01:42:26 crbMvm  9:    8    2
#> Jul08 01:42:26 crbMvm 10:    9    8
#> Jul08 01:42:26 crbMvm 11:   10    1
#> Jul08 01:42:26 frSprd  total elpsd: 1.7 secs | 1 fireSpread stats 5
#> Jul08 01:42:26 frSprd  1:    0   32
#> Jul08 01:42:26 frSprd  2:    1    3
#> Jul08 01:42:26 frSprd  3:    2    8
#> Jul08 01:42:26 frSprd  4:    3    1
#> Jul08 01:42:26 frSprd  5:    4   11
#> Jul08 01:42:26 frSprd  6:    5    4
#> Jul08 01:42:26 frSprd  7:    6    7
#> Jul08 01:42:26 frSprd  8:    7   23
#> Jul08 01:42:26 frSprd  9:    8    2
#> Jul08 01:42:26 frSprd 10:    9    8
#> Jul08 01:42:26 frSprd 11:   10    1
#> simList saved in
#>  SpaDES.core:::.pkgEnv$.sim 
#> It will be deleted at next spades() call.
#> Jul08 01:42:26       Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> simList saved in
#>  SpaDES.core:::.pkgEnv$.sim 
#> It will be deleted at next spades() call.
#> Setting:
#>   options(
#>     spades.modulePath = '/home/runner/work/_temp/Library/SpaDES.core/sampleModules'
#>   )
#> Paths set to:
#>   options(
#>     rasterTmpDir = '/tmp/Rtmp8X08T2/scratch/raster'
#>     reproducible.cachePath = '/tmp/Rtmp8X08T2/cache'
#>     spades.inputPath = '/tmp/Rtmp8X08T2/inputs'
#>     spades.outputPath = '/tmp/Rtmp8X08T2'
#>     spades.modulePath = '/home/runner/work/_temp/Library/SpaDES.core/sampleModules'
#>     spades.scratchPath = '/tmp/Rtmp8X08T2/scratch'
#>   )
#>   terra::terraOptions(tempdir = '/tmp/Rtmp8X08T2/scratch/terra'
#> Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> defineParameter: '.useCache' is not of specified type 'logical'.
#> The following .globals were used:
#>              module    global
#> 1:  caribouMovement stackName
#> 2:       fireSpread stackName
#> 3: randomLandscapes stackName
#> Jul08 01:42:29       Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> simList saved in
#>  SpaDES.core:::.pkgEnv$.sim 
#> It will be deleted at next spades() call.
#> Saving large object (cacheId: e796b390e194e61e) to Cache: 82.8 Mb
#>  Done!
#>    user  system elapsed 
#>   3.962   0.192   4.161 
#> Setting:
#>   options(
#>     spades.modulePath = '/home/runner/work/_temp/Library/SpaDES.core/sampleModules'
#>   )
#> Paths set to:
#>   options(
#>     rasterTmpDir = '/tmp/Rtmp8X08T2/scratch/raster'
#>     reproducible.cachePath = '/tmp/Rtmp8X08T2/cache'
#>     spades.inputPath = '/tmp/Rtmp8X08T2/inputs'
#>     spades.outputPath = '/tmp/Rtmp8X08T2'
#>     spades.modulePath = '/home/runner/work/_temp/Library/SpaDES.core/sampleModules'
#>     spades.scratchPath = '/tmp/Rtmp8X08T2/scratch'
#>   )
#>   terra::terraOptions(tempdir = '/tmp/Rtmp8X08T2/scratch/terra'
#> Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> defineParameter: '.useCache' is not of specified type 'logical'.
#> The following .globals were used:
#>              module    global
#> 1:  caribouMovement stackName
#> 2:       fireSpread stackName
#> 3: randomLandscapes stackName
#>   ...(Object to retrieve (e796b390e194e61e.rds))
#>      loaded cached result from previous spades call
#>    user  system elapsed 
#>   0.829   0.000   0.832 
#> Jul08 01:42:35       Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> simList saved in
#>  SpaDES.core:::.pkgEnv$.sim 
#> It will be deleted at next spades() call.
#> Jul08 01:42:35       NOTE: caribouMovement not specified in events argument. 
#> Jul08 01:42:35       This means all events in the module(s) will run. You may have intended to add e.g.,
#> Jul08 01:42:35       list(caribouMovement= 'init')
#> Jul08 01:42:35       Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> simList saved in
#>  SpaDES.core:::.pkgEnv$.sim 
#> It will be deleted at next spades() call.
#> Setting:
#>   options(
#>     spades.modulePath = '/home/runner/work/_temp/Library/SpaDES.core/sampleModules'
#>   )
#> Paths set to:
#>   options(
#>     rasterTmpDir = '/tmp/Rtmp8X08T2/scratch/raster'
#>     reproducible.cachePath = '/tmp/Rtmp8X08T2/cache'
#>     spades.inputPath = '/tmp/Rtmp8X08T2/inputs'
#>     spades.outputPath = '/tmp/Rtmp8X08T2'
#>     spades.modulePath = '/home/runner/work/_temp/Library/SpaDES.core/sampleModules'
#>     spades.scratchPath = '/tmp/Rtmp8X08T2/scratch'
#>   )
#>   terra::terraOptions(tempdir = '/tmp/Rtmp8X08T2/scratch/terra'
#> Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> defineParameter: '.useCache' is not of specified type 'logical'.
#> The following .globals were used:
#>              module    global
#> 1:  caribouMovement stackName
#> 2:       fireSpread stackName
#> 3: randomLandscapes stackName
#> Jul08 01:42:37       NOTE: caribouMovement not specified in events argument. 
#> Jul08 01:42:37       This means all events in the module(s) will run. You may have intended to add e.g.,
#> Jul08 01:42:37       list(caribouMovement= 'init')
#> Jul08 01:42:37       Using setDTthreads(1). To change: 'options(spades.DTthreads = X)'.
#> simList saved in
#>  SpaDES.core:::.pkgEnv$.sim 
#> It will be deleted at next spades() call.
# }